+++ src/main/java/tornadofx/ViewModel.kt
@@ -338,8 +338,24 @@ fun <V : ViewModel, T> V.rebindOnChange(observable: ObservableValue<T>, op: (V.(
 /**
  * Rebind the itemProperty of the ViewModel when the itemProperty in the ListCellFragment changes.
  */
-fun <V : ItemViewModel<T>, T> V.bindTo(listCellFragment: ListCellFragment<T>): V {
-    itemProperty.bind(listCellFragment.itemProperty)
+fun <V : ItemViewModel<T>, T> V.bindTo(cellFragment: ListCellFragment<T>): V {
+    itemProperty.bind(cellFragment.itemProperty)
+    return this
+}
+
+/**
+ * Rebind the itemProperty of the ViewModel when the itemProperty in the TableCellFragment changes.
+ */
+fun <V : ItemViewModel<T>, S, T> V.bindToItem(cellFragment: TableCellFragment<S, T>): V {
+    itemProperty.bind(cellFragment.itemProperty)
+    return this
+}
+
+/**
+ * Rebind the rowItemProperty of the ViewModel when the itemProperty in the TableCellFragment changes.
+ */
+fun <V : ItemViewModel<S>, S, T> V.bindToRowItem(cellFragment: TableCellFragment<S, T>): V {
+    itemProperty.bind(cellFragment.rowItemProperty)
     return this
 }
 
@@ -349,7 +365,7 @@ fun <V : ViewModel, T : ObservableValue<X>, X> V.dirtyStateFor(modelField: KProp
 }
 
 fun <V : ViewModel, T> V.rebindOnTreeItemChange(observable: ObservableValue<TreeItem<T>>, op: V.(T?) -> Unit) {
-    observable.addListener { observableValue, oldValue, newValue ->
+    observable.addListener { _, _, newValue ->
         op(newValue?.value)
         rollback()
     }